Outlining Vulkan
Alright, time for the next step in linux building. Vulkan seems like the best next step, not to mention the way to go in terms of rendering besides OpenGL. I have paths corrected and will figure out how to create windows for X11. Here is the expected window creation path for linux in the future:

Engine => Shader Compilation (if necessary) => VK Context => Window Creation

I expect to follow the same scheme as D3D11 to keep device handles rather than releasing them after usage so that window creation is faster. Compilation flags and runtime flags will determine which path to take in WindowConfig as well.

There is also a new "OS" definition inside WindowConfig so that paths that usually take GDI and D3D11 now take Vulkan for only Linux. All original paths for Windows should still work as normal.

+++ CS/Angene/Angene.Graphics/DX11/DX11Graphics.cs
@@ -29,7 +29,7 @@ namespace Angene.Graphics.DX11
         private bool _sharingDevice = false;
         public IntPtr ContextHandle => _context;
 
-        public nint Handle => (nint)_device;
+        public IntPtr Handle => _device;
 
         public DX11GraphicsContext(IntPtr hwnd, int width, int height, IntPtr existingDevice, IntPtr existingContext)
         {
